sosot/requirements.txt
mattermostdriver
langchain-core
langchain-chroma
langchain-community
langchain-ollama
langchain-text-splitters
chromadb
pypdf
python-dotenv
requests
pytest
unstructured
